2.4 release

+++ mupdf/source/fitz/ucdn.c
@@ -33,7 +33,7 @@ typedef struct {
 } MirrorPair;
 
 typedef struct {
-	unsigned int start;
+	int start;
 	short count, index;
 } Reindex;
 
@@ -91,9 +91,9 @@ static int get_comp_index(unsigned int code, const Reindex *idx)
 
 	for (i = 0; idx[i].start; i++) {
 		const Reindex *cur = &idx[i];
-		if (code < cur->start)
+		if (code < (unsigned)cur->start)
 			return -1;
-		if (code <= cur->start + cur->count) {
+		if (code <= (unsigned)cur->start + cur->count) {
 			return cur->index + (code - cur->start);
 		}
 	}
